trash/requirements.txt
grpcio==1.56.0
grpcio-tools==1.56.0
protobuf==4.23.4
rocksdb==0.7.0
asyncio==3.4.3